Issue #6
This was a pretty major change to the plugin. In order to bring this
plugins functionality more in line with Jenkins idioms, I changed the
withPythonEnv step to apply environmental variable changes skimmed from
the activiation of the virtualenv. This is how virtualenv works under
the hood when active anyways.

As such, we no longer need the pybat and pysh steps, the regular sh
and bat steps work as intended. These have been removed as part of this
commit